% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dockerfile_with_renv_custom.R
\name{add_dockerfile_with_renv_custom}
\alias{add_dockerfile_with_renv_custom}
\title{add_dockerfile_with_renv_custom}
\usage{
add_dockerfile_with_renv_custom(gcp_project, ...)
}
\arguments{
\item{gcp_project}{name of GCP project}

\item{...}{argument to pass to golem::add_dockerfile_with_renv_custom()}
}
\value{
Creates or replace deploy/ folder with corresponding resources
}
\description{
Custom wrapper over golem::add_dockerfile_with_renv() that insludes .Renviron lines and .secrets in Dockerfile
}
