% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_wrappers.R
\name{bq_get}
\alias{bq_get}
\title{bq_get}
\usage{
bq_get(query, dataset = NULL, location = NULL)
}
\arguments{
\item{query}{SQL query to execute}

\item{dataset}{str name of the dataset that will perform the job}

\item{location}{str name of the wkt column to convert to sf}
}
\value{
invisible.
}
\description{
wrapper helpers to communicate with Bigquery
}
