% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_places.R
\name{get_google_places}
\alias{get_google_places}
\alias{get_google_places_pagination}
\title{get_google_places}
\usage{
get_google_places(...)

get_google_places_pagination(...)
}
\arguments{
\item{...}{parameters to pass to googleway::get_places}
}
\value{
get_google_places returns tibble. get_google_places_pagination returns response list
}
\description{
Wrapper of googleway::get_places to paginate and get all possible results.
}
\details{
get_google_places cleans the response to get a tibble of places and attributes:
name, place_id, types, vicinity, rating, user_ratings_total, lat, lng
get_google_places_pagination returns a list with max 3 responses (Google Places API return max 3 resposes of 20 places each)
}
