% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{mod_chat_ui}
\alias{mod_chat_ui}
\alias{mod_chat_server}
\title{chat UI Function}
\usage{
mod_chat_ui(id, height = "600px")

mod_chat_server(id, user_id, assistant_id, project, dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{user_id}{Vikua User ID to log the conversation}

\item{assistant_id}{OpenAI assistant ID to use in the conversation}

\item{project, dataset}{path to the project and dataset in BigQuery}
}
\description{
Idiomatic conversation with VikuaBot (OpenAI assistants)
}
