% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geocoding.R
\name{mod_geocoding_ui}
\alias{mod_geocoding_ui}
\alias{mod_geocoding_server}
\title{geocoding UI Function}
\usage{
mod_geocoding_ui(id)

mod_geocoding_server(id, example_dataset, package_name)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{example_dataset}{str. name of the .csv file to be read and display via DT::datatable}

\item{package_name}{str. name of the shiny-package}
}
\description{
A shiny Module that allows to geocode addresses.
The user `fileInput`s a .csv file with the column *address* and press the bttn to geocode via geocode_apply.
A leaflet map displays with the marker locations.
If any address can`t be geocoded, a DT::datatable displays below the map
User can manually locate addresses or modify them in the map via mod_geocoding_manual or by dragging markers
User can download the same fileInput .csv with new columns lat, lng and place_id
}
