% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_download}
\alias{name_download}
\alias{column_card}
\alias{column_card_button}
\alias{expand_air_month_picker_output}
\alias{compile_chat}
\title{utils}
\usage{
name_download(file_name, module_name)

column_card(width, title, ...)

column_card_button(width, title, button, ...)

expand_air_month_picker_output(output_vector)

compile_chat(i, data)
}
\arguments{
\item{file_name}{str}

\item{module_name}{str}

\item{width}{passed to bs4Dash::column}

\item{title}{str Title for the output}

\item{...}{output to show in column}

\item{button}{button widget to show next to title}

\item{output_vector}{output of shinyWidget::AirDatePicker}
}
\description{
Utils
}
