% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_app_vk.R
\name{secure_app_vk}
\alias{secure_app_vk}
\alias{check_credentials_vk}
\alias{log_store_custom_vk}
\title{secure_app_vk}
\usage{
secure_app_vk(app_ui, prod = T, tags_vk)

check_credentials_vk(
  prod = T,
  gcp_project,
  bq_dataset,
  bq_table = "credentials",
  fields = c("permission", "company")
)

log_store_custom_vk(app_name, gcp_project, bq_dataset, bq_table, prod)
}
\arguments{
\item{app_ui}{function application user-interface}

\item{prod}{boolean identifying if the app communicates with BQ to ask for credentials and save logs}

\item{tags_vk}{named list with the top, bottom and img strings to include in login module}

\item{gcp_project, bq_dataset, bq_table}{string names of GCP resources to retrieve and write}

\item{fields}{string vector of fields to return from `bq_table`. These will be available in auth reactive}

\item{app_name}{string to include in logs}
}
\value{
invisible.
}
\description{
login module for vikua shiny apps
}
