% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyvktools-package.R
\docType{package}
\name{shinyvktools-package}
\alias{shinyvktools}
\alias{shinyvktools-package}
\title{shinyvktools: Utils for Vikua Shinyapps}
\description{
Vikua is a urban-tech latam startup focused on provide urban-oriented analytic solutions This package groups common functions between the apps.
}
\author{
\strong{Maintainer}: Rodrigo Serrano \email{roserrano@vikua.com}

}
\keyword{internal}
